+++ back/prisma/migrations/20240122155001_remove_status/migration.sql
@@ -1,0 +1,11 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `status` on the `user` table. All the data in the column will be lost.
+
+*/
+-- AlterTable
+ALTER TABLE "user" DROP COLUMN "status";
+
+-- DropEnum
+DROP TYPE "USER_STATUS";