+++ sso/src/schemas/user/userSchema.ts
@@ -6,12 +6,17 @@ export enum UserRole {
   MENTOR = 'MENTOR',
   REGISTERED = 'REGISTERED',
 }
+export enum UserStatus {
+  ACTIVE = 'ACTIVE',
+  INACTIVE = 'INACTIVE',
+}
 export const userSchema = z.object({
   id: z.string().cuid2(),
   dni: dniSchema,
   email: z.string().email().nonempty().openapi({ example: 'user@example.cat' }),
   password: z.string().min(8),
   role: z.nativeEnum(UserRole),
+  status: z.nativeEnum(UserStatus),
   createdAt: z.string().datetime(),
   updatedAt: z.string().datetime(),
   itineraryId: z.string().cuid2(),